+++ lib/Elastica/QueryBuilder/DSL/Filter.php
@@ -62,12 +62,9 @@ class Filter implements DSL
      *
      * @return BoolAnd
      */
-    public function bool_and(array $filters)
+    public function bool_and(array $filters = array())
     {
-        $and = new BoolAnd();
-        $and->setFilters($filters);
-
-        return $and;
+        return new BoolAnd($filters);
     }
 
     /**
@@ -101,14 +98,14 @@ class Filter implements DSL
      *
      * @link http://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-geo-bounding-box-filter.html
      *
-     * @param string $field
+     * @param string $key
      * @param array  $coordinates
      *
      * @return GeoBoundingBox
      */
-    public function geo_bounding_box($field, array $coordinates)
+    public function geo_bounding_box($key, array $coordinates)
     {
-        return new GeoBoundingBox($field, $coordinates);
+        return new GeoBoundingBox($key, $coordinates);
     }
 
     /**
@@ -197,16 +194,16 @@ class Filter implements DSL
      *
      * @link http://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-geohash-cell-filter.html
      *
-     * @param string       $field     The field on which to filter
+     * @param string       $key       The field on which to filter
      * @param array|string $location  Location as coordinates array or geohash string ['lat' => 40.3, 'lon' => 45.2]
      * @param int|string   $precision length of geohash prefix or distance (3, or "50m")
      * @param bool         $neighbors If true, filters cells next to the given cell.
      *
      * @return GeohashCell
      */
-    public function geohash_cell($field, $location, $precision = -1, $neighbors = false)
+    public function geohash_cell($key, $location, $precision = -1, $neighbors = false)
     {
-        return new GeohashCell($field, $location, $precision, $neighbors);
+        return new GeohashCell($key, $location, $precision, $neighbors);
     }
 
     /**
@@ -219,7 +216,7 @@ class Filter implements DSL
      *
      * @return HasChild
      */
-    public function has_child($query, $type)
+    public function has_child($query, $type = null)
     {
         return new HasChild($query, $type);
     }
@@ -249,7 +246,7 @@ class Filter implements DSL
      *
      * @return Ids
      */
-    public function ids($type, array $ids)
+    public function ids($type = null, array $ids = array())
     {
         return new Ids($type, $ids);
     }
@@ -304,7 +301,7 @@ class Filter implements DSL
      *
      * @return Missing
      */
-    public function missing($field)
+    public function missing($field = '')
     {
         return new Missing($field);
     }
@@ -340,11 +337,14 @@ class Filter implements DSL
      *
      * @link http://www.elastic.co/guide/en/elasticsearch/reference/0.90/query-dsl-numeric-range-filter.html
      *
+     * @param string $fieldName Field name
+     * @param array  $args      Field arguments
+     *
      * @return NumericRange
      */
-    public function numeric_range()
+    public function numeric_range($fieldName = '', array $args = array())
     {
-        return new NumericRange();
+        return new NumericRange($fieldName, $args);
     }
 
     /**
@@ -356,12 +356,9 @@ class Filter implements DSL
      *
      * @return BoolOr
      */
-    public function bool_or($filters)
+    public function bool_or(array $filters = array())
     {
-        $or = new BoolOr();
-        $or->setFilters($filters);
-
-        return $or;
+        return new BoolOr($filters);
     }
 
     /**
@@ -374,7 +371,7 @@ class Filter implements DSL
      *
      * @return Prefix
      */
-    public function prefix($field, $prefix)
+    public function prefix($field = '', $prefix = '')
     {
         return new Prefix($field, $prefix);
     }
@@ -384,11 +381,11 @@ class Filter implements DSL
      *
      * @link http://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-filter.html
      *
-     * @param AbstractQuery $query
+     * @param array|AbstractQuery $query
      *
      * @return QueryFilter
      */
-    public function query(AbstractQuery $query)
+    public function query($query = null)
     {
         return new QueryFilter($query);
     }
@@ -403,7 +400,7 @@ class Filter implements DSL
      *
      * @return Range
      */
-    public function range($fieldName, array $args)
+    public function range($fieldName = '', array $args = array())
     {
         return new Range($fieldName, $args);
     }
@@ -413,14 +410,15 @@ class Filter implements DSL
      *
      * @link http://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-regexp-filter.html
      *
-     * @param string $field
-     * @param string $regexp
+     * @param string $field   Field name
+     * @param string $regexp  Regular expression
+     * @param array  $options Regular expression options
      *
      * @return Regexp
      */
-    public function regexp($field, $regexp)
+    public function regexp($field = '', $regexp = '', $options = array())
     {
-        return new Regexp($field, $regexp);
+        return new Regexp($field, $regexp, $options);
     }
 
     /**
@@ -432,7 +430,7 @@ class Filter implements DSL
      *
      * @return Script
      */
-    public function script($script)
+    public function script($script = null)
     {
         return new Script($script);
     }
@@ -456,14 +454,14 @@ class Filter implements DSL
      *
      * @link http://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-terms-filter.html
      *
-     * @param string $field
+     * @param string $key
      * @param array  $terms
      *
      * @return Terms
      */
-    public function terms($field, array $terms)
+    public function terms($key = '', array $terms = array())
     {
-        return new Terms($field, $terms);
+        return new Terms($key, $terms);
     }
 
     /**
@@ -475,7 +473,7 @@ class Filter implements DSL
      *
      * @return Type
      */
-    public function type($type)
+    public function type($type = null)
     {
         return new Type($type);
     }